Revert "fix: meeting note dto id field 타입 수정 (#56)"
This reverts commit 14abcc34c3e5049500f7c600b6063e32b8a56d61.

+++ src/main/java/com/ellu/looper/dto/MeetingNoteRequest.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MeetingNoteRequest {
-  private int project_id;
+  private Long project_id;
   private String content;
   private String nickname;
   private String position;
